requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
Pillow==11.0.0
numpy==2.1.3
minio==7.2.12
supabase==2.10.0
joblib==1.4.2
scikit-learn==1.5.2
tensorflow-cpu==2.16.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
